Move caret related class into edit namespace
