+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/CaretLayer.cs
@@ -12,7 +12,7 @@ using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Lyrics;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition.Algorithms;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Carets;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit.Carets;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Utils;
