+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/Carets/DrawableLyricSplitterCaret.cs
@@ -10,9 +10,10 @@ using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Carets
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit.Carets
 {
     public class DrawableLyricSplitterCaret : DrawableLyricTextCaret
     {
