+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/TimeTagLayer.cs
@@ -7,7 +7,7 @@ using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Parts;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit.Components;
 using osu.Game.Rulesets.Karaoke.Objects;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit
